requirements.txt
google-generativeai
python-dotenv
anthropic
backoff
openai
transformers
torch
together